+++ src/networkx_temporal/classes/multidigraph.py
@@ -2,10 +2,10 @@ from typing import Optional
 
 import networkx as nx
 
-from .base import TemporalBase
+from .abc import TemporalABC
 
 
-class TemporalMultiDiGraph(TemporalBase, nx.MultiDiGraph if nx.__version__ >= "2.8.1" else object):
+class TemporalMultiDiGraph(TemporalABC, nx.MultiDiGraph if nx.__version__ >= "2.8.1" else object):
     """
     Creates a temporal directed multigraph. Allows parallel edges among the same pair of nodes.
 
@@ -15,12 +15,12 @@ class TemporalMultiDiGraph(TemporalBase, nx.MultiDiGraph if nx.__version__ >= "2
     :func:`neighbors`, :func:`subgraph`, :func:`to_directed`, and :func:`to_undirected`,
     as well as additional methods implemented for handling temporal graphs and snapshots.
 
-    This is equivalent to calling :func:`~networkx_temporal.graph.temporal_graph` with ``directed=True``
+    This is equivalent to calling :func:`~networkx_temporal.classes.temporal_graph` with ``directed=True``
     and ``multigraph=True``.
 
     .. seealso::
 
-       The :class:`~networkx_temporal.graph.TemporalGraph` class documentation
+       The :class:`~networkx_temporal.classes.TemporalGraph` class documentation
        for details on its implemented methods.
 
     :param int t: Number of temporal graphs to initialize. Optional. Default is ``1``.
@@ -28,4 +28,4 @@ class TemporalMultiDiGraph(TemporalBase, nx.MultiDiGraph if nx.__version__ >= "2
     :note: Documentation on inherited methods is available only if ``networkx>=2.8.1``.
     """
     def __init__(self, t: Optional[int] = None):
-        super().__init__(t=t, directed=True, multigraph=True)
+        super().__init__(t=t, create_using=nx.MultiDiGraph)