+++ src/networkx_temporal/utils/convert/convert.py
@@ -63,8 +63,8 @@ def convert(G: Union[TemporalGraph, StaticGraph, list], to: FORMATS, *args, **kw
 
     Convert the `Karate Club
     <https://networkx.org/documentation/stable/auto_examples/graph/plot_karate_club.html>`__
-    graph dataset from NetworkX into a PyTorch Geometric `Data
-    <https://pytorch-geometric.readthedocs.io/en/latest/generated/torch_geometric.data.Data.html>`__
+    dataset from NetworkX into a PyTorch Geometric `Data
+    <https://pytorch-geometric.readthedocs.readwrite/en/latest/generated/torch_geometric.data.Data.html>`__
     object:
 
     .. code-block:: python
@@ -81,7 +81,7 @@ def convert(G: Union[TemporalGraph, StaticGraph, list], to: FORMATS, *args, **kw
         Graph named "Zachary's Karate Club" with 34 nodes and 78 edges
         Data(edge_index=[2, 156], club=[34], weight=[156], name='Zachary's Karate Club', num_nodes=34)
 
-    :param object G: Graph object. Accepts a :class:`~networkx_temporal.graph.TemporalGraph`, a
+    :param object G: Graph object. Accepts a :class:`~networkx_temporal.classes.TemporalGraph`, a
         single static NetworkX graph, or a list of static NetworkX graphs as input.
     :param str to: Package name or alias to convert the graph object.
     :param args: Additional positional arguments for the conversion function.
@@ -89,7 +89,7 @@ def convert(G: Union[TemporalGraph, StaticGraph, list], to: FORMATS, *args, **kw
 
     :rtype: Any
 
-    :note: Available both as a function and as a method from :class:`~networkx_temporal.graph.TemporalGraph` objects.
+    :note: Available both as a function and as a method from :class:`~networkx_temporal.classes.TemporalGraph` objects.
     """
     pkg = ALIASES.get(to, to)
     func = ALIASES.get(pkg, pkg)