+++ src/networkx_temporal/utils/convert/torch_geometric.py
@@ -1,14 +1,14 @@
 from typing import Union
 
-from ..networkx import is_static_graph, is_temporal_graph
+from ..utils import is_static_graph, is_temporal_graph
 from ...typing import StaticGraph, TemporalGraph
 
 
 def to_torch_geometric(G: Union[TemporalGraph, StaticGraph, list], *args, **kwargs):
     """
-    Convert from NetworkX to `PyTorch Geometric <https://pytorch-geometric.readthedocs.io/>`__.
+    Convert from NetworkX to `PyTorch Geometric <https://pytorch-geometric.readthedocs.readwrite/>`__.
 
-    :param object G: Graph object. Accepts a :class:`~networkx_temporal.graph.TemporalGraph`, a
+    :param object G: Graph object. Accepts a :class:`~networkx_temporal.classes.TemporalGraph`, a
         single static NetworkX graph, or a list of static NetworkX graphs as input.
     :param args: Positional arguments.
     :param kwargs: Keyword arguments.
@@ -17,7 +17,7 @@ def to_torch_geometric(G: Union[TemporalGraph, StaticGraph, list], *args, **kwar
 
     :note: Wrapper function for
         `torch_geometric.utils.from_networkx
-        <https://pytorch-geometric.readthedocs.io/en/stable/modules/utils.html#torch_geometric.utils.from_networkx>`__.
+        <https://pytorch-geometric.readthedocs.readwrite/en/stable/modules/utils.html#torch_geometric.utils.from_networkx>`__.
     """
     import torch_geometric as pyg